Major refactor: split plotting into two scripts
